Workspace: normalize filename on diagnostics notification
On diagnostics notification there is the `-` file that was not being
translated to the real filename when selecting the diagnostics results from
a %hash<filename => diag>, causing a notification with erroneous diagnostics
(empty, as no errors existed for filename `-`), overriding any other
notification for that same file depending on the order of delivery to the
client.

This patch adds just a simple reassignment to $filename using the $uri as
basis, since it was correct even for the `-` filename.